requirements.txt
anyio==4.3.0
asyncio==3.4.3
certifi==2024.2.2
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
pyaes==1.6.1
pyasn1==0.5.1
python-telegram-bot==20.8
rsa==4.9
sniffio==1.3.0
Telethon==1.34.0
